fix(schema): Correct schema definitions
Various schemas have been updated for correctness.

Notably, although an incorrect Elem definition for custom_headers is
provided, this field still does not work as is due to a limitation in
Terraform.  Maps of mixed types are not possible to use or define, and
therefore a workaround would be required to implement this field.  This
more correct definition is here in case this Terraform behavior changes
in the future.