Support building structs from list resources
targetSipCredentials schema validation was broken by a combination of
newer version of Terraform with the new plugin SDK.
Newer versions of Terraform do not permit declaring map schemas as
resources, meaning that there is no way to validate maps with elements
of disparate types or to validate specific key names in a map.  The
alternatives are to perform schema validtion manually elsewhere, or to
declare them as resources.  Since we already fall back to
resource declarations for lists of maps, repeating the practice seemed
like the least worst option.  Schema validation for this case is done by
declaring the type as a list with a maximum length of 1, and elements
declared as resources.