BREAKING CHANGE: labels/groups attachment to tests/agents updates (#125)
* BREAKING CHANGE: labels/groups attachment to tests/agents updates

Going forward it will no longer be possible to attach a label/group
to a test in any test resource. Test resources obtain the labels
linked to them via the API automatically (computed fields). Labels
are now linked to tests only in the label resource. This diverges
from what the ThousandEyes API allows, but is necessary considering
how terraform works, as it was previously possible to have a test
resource cancel out a label that is being added to it in a label
resource, because the test resource wasn't setting it.

* docs: fix account name description grammar